Fix compiler warnings new in VS2015
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@26611 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNStatusListCtrl.cpp
@@ -6792,9 +6792,9 @@ BOOL CSVNStatusListCtrl::OnWndMsg(UINT message, WPARAM wParam, LPARAM lParam, LR
                     int succeededItems = 0;
                     PIDLIST_RELATIVE pidl = NULL;
 
-                    size_t bufsize = 1024;
+                    int bufsize = 1024;
                     std::unique_ptr<WCHAR[]> filepath(new WCHAR[bufsize]);
-                    for (size_t i = 0; i < nItems; i++)
+                    for (int i = 0; i < nItems; i++)
                     {
                         if (bufsize < targetList[i].GetWinPathString().GetLength())
                         {
